Move domain model to models and remove unused model
